other types of allocator added
not tested, basic types renamed also

+++ include/GC/Utils.h
@@ -9,22 +9,22 @@
 #define GC_GET_WORD_SIZE (::gc::priv::size_if<std::numeric_limits<size_t>::max() == 0xffffffff, 4, 8>::value)
 namespace gc
 {
-	typedef int8_t int8;
-	typedef uint8_t uint8;
-	typedef int16_t int16;
-	typedef uint16_t uint16;
-	typedef int32_t int32;
-	typedef uint32_t uint32;
-	typedef int64_t int64;
-	typedef uint64_t uint64;
+	typedef int8_t i8;
+	typedef uint8_t u8;
+	typedef int16_t i16;
+	typedef uint16_t u16;
+	typedef int32_t i32;
+	typedef uint32_t u32;
+	typedef int64_t i64;
+	typedef uint64_t u64;
 	const float Pi = 3.14159265358979323846f;
 	
 	template<class T>
 	struct TypeName
 	{
 		inline static std::string get(){
-			static const unsigned int FRONT_SIZE = sizeof("gc::TypeName<") - 1u;
-			static const unsigned int BACK_SIZE = sizeof(">::get") - 1u;
+			static const size_t FRONT_SIZE = sizeof("gc::TypeName<") - 1u;
+			static const size_t BACK_SIZE = sizeof(">::get") - 1u;
 			static const size_t size = sizeof(__FUNCTION__) - FRONT_SIZE - BACK_SIZE;
 			return std::string(__FUNCTION__ + FRONT_SIZE, size - 1u);
 		}
@@ -65,11 +65,11 @@ namespace gc
 		};
 	}
 }
-typedef int8_t int8;
-typedef uint8_t uint8;
-typedef int16_t int16;
-typedef uint16_t uint16;
-typedef int32_t int32;
-typedef uint32_t uint32;
-typedef int64_t int64;
-typedef uint64_t uint64;
+typedef int8_t i8;
+typedef uint8_t u8;
+typedef int16_t i16;
+typedef uint16_t u16;
+typedef int32_t i32;
+typedef uint32_t u32;
+typedef int64_t i64;
+typedef uint64_t u64;
